Include headers and modulemap
Previously we were still relying on the system toolchain's headers /
modulemap for this library which isn't valid at the point that changes
and SwiftSyntax adopts the new functionality. I first hit this trying to
build Swift 5.7's SwiftSyntax with Xcode 13.4. Theoretically things
could be entirely incompatible and it would require the new Xcode
(because of a new Swift language feature), but this is better than
nothing for these specific cases.